Adding HybSel a Hybrid feature selection method using prior knowledge
This method utilizes data from the PDR GTEx Portal and GWAS Catalog to create a knowledge matrix. It employs the F1 Score as a performance metric to validate the optimal weight vector. Validation of the F1 score involves assigning reported genes from the GWAS catalog for the chosen phenotype (PD).